Update contributing.md
Apart from editing. there are two comments:
In the Code Generation section there's a bullet, is there a plan to add content there? In any case no point in publishing this page with an empty bullet. 

In Missing a Feature? I don't understand what you wanted to say here "please submit an issue with a proposal for your work first, to be sure that we can use it. Please consider what kind of change it is:" After you clarify will edit it.